Compress cubes to 4 bytes each, do a bit of work for tiny speedup of worldgen

+++ Rendering/Cube.hpp
@@ -1,9 +1,11 @@
 #ifndef CUBE_H
 #define CUBE_H
 
+#include <cstdint>
+
 struct Cube {
-  float x, y, z;
-  int type;
+  uint8_t x, y, z;
+  uint8_t type;
 };
 
 #endif
